sprs-ai-portfolio-system/backend/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
pymongo==4.8.0
pandas==2.2.2
scikit-learn==1.5.1
numpy==1.26.4
PuLP==2.9.0
python-dotenv==1.0.1
